requirements.txt
fastapi==0.97.0
Jinja2==3.1.2
openai==0.27.8
Pillow==9.5.0
python-dotenv==1.0.0
requests==2.31.0
uvicorn==0.22.0
